extracted reused methods
