+++ test/unit/specs/components/DecisioningEngine/contextTestUtils.js
@@ -1,0 +1,124 @@
+import createContextProvider from "../../../../../src/components/DecisioningEngine/createContextProvider";
+import createOnResponseHandler from "../../../../../src/components/DecisioningEngine/createOnResponseHandler";
+import createEventRegistry from "../../../../../src/components/DecisioningEngine/createEventRegistry";
+import createDecisionProvider from "../../../../../src/components/DecisioningEngine/createDecisionProvider";
+
+export const proposition = {
+  id: "2e4c7b28-b3e7-4d5b-ae6a-9ab0b44af87e",
+  items: [
+    {
+      schema: "https://ns.adobe.com/personalization/mock-action",
+      data: {
+        hello: "kitty"
+      },
+      id: "79129ecf-6430-4fbd-955a-b4f1dfdaa6fe"
+    }
+  ],
+  scope: "web://mywebsite.com"
+};
+
+export const mockWindow = ({
+  title = "My awesome website",
+  referrer = "https://www.google.com/search?q=adobe+journey+optimizer&oq=adobe+journey+optimizer",
+  url = "https://pro.mywebsite.org:8080/about?m=1&t=5&name=jimmy#home",
+  width = 100,
+  height = 100,
+  scrollX = 0,
+  scrollY = 10,
+  userAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/78.0.3904.87 Safari/537.36"
+}) => ({
+  title,
+  referrer,
+  url,
+  width,
+  height,
+  scrollX,
+  scrollY,
+  navigator: {
+    userAgent
+  }
+});
+
+export const payloadWithCondition = condition => {
+  return {
+    id: "2e4c7b28-b3e7-4d5b-ae6a-9ab0b44af87e",
+    items: [
+      {
+        id: "79129ecf-6430-4fbd-955a-b4f1dfdaa6fe",
+        schema: "https://ns.adobe.com/personalization/json-ruleset-item",
+        data: {
+          content: JSON.stringify({
+            version: 1,
+            rules: [
+              {
+                condition: {
+                  definition: {
+                    conditions: [condition],
+                    logic: "and"
+                  },
+                  type: "group"
+                },
+                consequences: [
+                  {
+                    type: "item",
+                    detail: {
+                      schema:
+                        "https://ns.adobe.com/personalization/mock-action",
+                      data: {
+                        hello: "kitty"
+                      },
+                      id: "79129ecf-6430-4fbd-955a-b4f1dfdaa6fe"
+                    },
+                    id: "79129ecf-6430-4fbd-955a-b4f1dfdaa6fe"
+                  }
+                ]
+              }
+            ]
+          })
+        }
+      }
+    ],
+    scope: "web://mywebsite.com"
+  };
+};
+export const mockRulesetResponseWithCondition = condition => {
+  return {
+    getPayloadsByType: () => [
+      payloadWithCondition({
+        definition: {
+          conditions: [condition],
+          logic: "and"
+        },
+        type: "group"
+      })
+    ]
+  };
+};
+
+const mockEvent = { getContent: () => ({}), getViewName: () => undefined };
+
+export const setupResponseHandler = (applyResponse, window, condition) => {
+  const storage = jasmine.createSpyObj("storage", [
+    "getItem",
+    "setItem",
+    "clear"
+  ]);
+  const eventRegistry = createEventRegistry({ storage });
+  const decisionProvider = createDecisionProvider();
+
+  const contextProvider = createContextProvider({
+    eventRegistry,
+    window
+  });
+
+  const onResponseHandler = createOnResponseHandler({
+    decisionProvider,
+    applyResponse,
+    event: mockEvent,
+    decisionContext: contextProvider.getContext()
+  });
+
+  onResponseHandler({
+    response: mockRulesetResponseWithCondition(condition)
+  });
+};
