+++ test/unit/specs/components/DecisioningEngine/decisioningContext.timestamp.spec.js
@@ -1,133 +1,13 @@
-import createOnResponseHandler from "../../../../../src/components/DecisioningEngine/createOnResponseHandler";
-import createDecisionProvider from "../../../../../src/components/DecisioningEngine/createDecisionProvider";
-import createContextProvider from "../../../../../src/components/DecisioningEngine/createContextProvider";
-import createEventRegistry from "../../../../../src/components/DecisioningEngine/createEventRegistry";
-
-const mockWindow = ({
-  title = "My awesome website",
-  referrer = "https://www.google.com/search?q=adobe+journey+optimizer&oq=adobe+journey+optimizer",
-  url = "https://pro.mywebsite.org:8080/about?m=1&t=5&name=jimmy#home",
-  width = 100,
-  height = 100,
-  scrollX = 0,
-  scrollY = 10,
-  userAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/78.0.3904.87 Safari/537.36"
-}) => ({
-  title,
-  referrer,
-  url,
-  width,
-  height,
-  scrollX,
-  scrollY,
-  navigator: {
-    userAgent
-  }
-});
-
-const proposition = {
-  id: "2e4c7b28-b3e7-4d5b-ae6a-9ab0b44af87e",
-  items: [
-    {
-      schema: "https://ns.adobe.com/personalization/mock-action",
-      data: {
-        hello: "kitty"
-      },
-      id: "79129ecf-6430-4fbd-955a-b4f1dfdaa6fe"
-    }
-  ],
-  scope: "web://mywebsite.com"
-};
-
-const payloadWithCondition = condition => {
-  return {
-    id: "2e4c7b28-b3e7-4d5b-ae6a-9ab0b44af87e",
-    items: [
-      {
-        id: "79129ecf-6430-4fbd-955a-b4f1dfdaa6fe",
-        schema: "https://ns.adobe.com/personalization/json-ruleset-item",
-        data: {
-          content: JSON.stringify({
-            version: 1,
-            rules: [
-              {
-                condition: {
-                  definition: {
-                    conditions: [condition],
-                    logic: "and"
-                  },
-                  type: "group"
-                },
-                consequences: [
-                  {
-                    type: "item",
-                    detail: {
-                      schema:
-                        "https://ns.adobe.com/personalization/mock-action",
-                      data: {
-                        hello: "kitty"
-                      },
-                      id: "79129ecf-6430-4fbd-955a-b4f1dfdaa6fe"
-                    },
-                    id: "79129ecf-6430-4fbd-955a-b4f1dfdaa6fe"
-                  }
-                ]
-              }
-            ]
-          })
-        }
-      }
-    ],
-    scope: "web://mywebsite.com"
-  };
-};
-
-const mockRulesetResponseWithCondition = condition => {
-  return {
-    getPayloadsByType: () => [
-      payloadWithCondition({
-        definition: {
-          conditions: [condition],
-          logic: "and"
-        },
-        type: "group"
-      })
-    ]
-  };
-};
+import {
+  mockWindow,
+  setupResponseHandler,
+  proposition
+} from "./contextTestUtils";
 
 let mockedTimestamp;
 describe("DecisioningEngine:globalContext:timeContext", () => {
-  let storage;
-  let eventRegistry;
-  let decisionProvider;
-  let onResponseHandler;
   let applyResponse;
-
-  const mockEvent = { getContent: () => ({}), getViewName: () => undefined };
-
-  const setupResponseHandler = (window, condition) => {
-    const contextProvider = createContextProvider({
-      eventRegistry,
-      window
-    });
-
-    onResponseHandler = createOnResponseHandler({
-      decisionProvider,
-      applyResponse,
-      event: mockEvent,
-      decisionContext: contextProvider.getContext()
-    });
-
-    onResponseHandler({
-      response: mockRulesetResponseWithCondition(condition)
-    });
-  };
-
   beforeEach(() => {
-    storage = jasmine.createSpyObj("storage", ["getItem", "setItem", "clear"]);
-    eventRegistry = createEventRegistry({ storage });
-    decisionProvider = createDecisionProvider();
     applyResponse = jasmine.createSpy();
     mockedTimestamp = new Date(Date.UTC(2023, 4, 11, 13, 34, 56));
     jasmine.clock().install();
@@ -138,7 +18,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("satisfies rule based on matched pageLoadTimestamp", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "pageLoadTimestamp",
         matcher: "eq",
@@ -155,7 +35,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("does not satisfy rule due to unmatched pageLoadTimestamp", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "pageLoadTimestamp",
         matcher: "eq",
@@ -172,7 +52,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("satisfies rule based on matched currentTimestamp", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentTimestamp",
         matcher: "eq",
@@ -189,7 +69,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("does not satisfy rule due to unmatched currentTimestamp", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentTimestamp",
         matcher: "eq",
@@ -206,7 +86,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("satisfies rule based on matched currentDate", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentDate",
         matcher: "eq",
@@ -223,7 +103,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("does not satisfy rule due to unmatched currentDate", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentDate",
         matcher: "eq",
@@ -240,7 +120,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("satisfies rule based on matched currentDay", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentDay",
         matcher: "eq",
@@ -257,7 +137,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("does not satisfy rule due to unmatched currentDay", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentDay",
         matcher: "eq",
@@ -274,7 +154,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("satisfies rule based on matched currentHour", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentHour",
         matcher: "eq",
@@ -291,7 +171,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("does not satisfy rule due to unmatched currentHour", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentHour",
         matcher: "eq",
@@ -308,7 +188,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("satisfies rule based on matched currentMinute", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentMinute",
         matcher: "eq",
@@ -325,7 +205,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("does not satisfy rule due to unmatched currentMinute", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentMinute",
         matcher: "eq",
@@ -342,7 +222,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("satisfies rule based on matched currentMonth", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentMonth",
         matcher: "eq",
@@ -359,7 +239,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("does not satisfy rule due to unmatched currentMonth", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentMonth",
         matcher: "eq",
@@ -376,7 +256,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("satisfies rule based on matched currentYear", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentYear",
         matcher: "eq",
@@ -393,7 +273,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("does not satisfy rule due to unmatched currentYear", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "currentYear",
         matcher: "eq",
@@ -410,7 +290,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("satisfies rule based on matched pageVisitDuration", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "pageVisitDuration",
         matcher: "eq",
@@ -426,7 +306,7 @@ describe("DecisioningEngine:globalContext:timeContext", () => {
   });
 
   it("does not satisfy rule due to unmatched pageVisitDuration", () => {
-    setupResponseHandler(mockWindow({}), {
+    setupResponseHandler(applyResponse, mockWindow({}), {
       definition: {
         key: "pageVisitDuration",
         matcher: "eq",
